Refactor project structure FTW Gave some love to the types Add usefule doc for when zscaler zscales as usual...
