+++ .vscode/extensions.json
@@ -1,0 +1,5 @@
+{
+  "recommendations": [
+    "denoland.vscode-deno"
+  ]
+}
