+++ README.md
@@ -2,6 +2,43 @@
 
 `xmas` for short 🎄
 
+### Maintainers
+> **Setup**: After cloning, run `deno install` to install and cache all dependencies.
+
+> **VS Code Users**: When you first open this project, VS Code will suggest installing the Deno extension. Accept this suggestion to get proper TypeScript support, formatting, and linting as configured in the [.vscode/settings.json](.vscode/settings.json) file.
+
+> **Corporate Networks**: If you're behind a corporate firewall (like Zscaler), you may encounter SSL certificate issues when downloading dependencies. Use the configured tasks instead of direct Deno commands:
+
+**Development Commands**:
+- `deno task test` - Run all unit tests (handles corporate certificates)
+- `deno task cache` - Cache dependencies (handles corporate certificates)
+- `deno run --allow-net sandbox/index.ts` - Run sandbox for quick prototyping
+
+**Alternative Commands** (if not behind corporate firewall):
+- `deno test` - Run all unit tests
+- `deno cache --reload src/**/*.ts` - Cache dependencies
+
+### Troubleshooting
+
+**SSL Certificate Issues**:
+If you encounter errors like `invalid peer certificate: UnknownIssuer` when running Deno commands, you're likely behind a corporate firewall that intercepts SSL certificates.
+
+**Solution**: Use the configured tasks which include `DENO_TLS_CA_STORE=system`:
+```bash
+deno task test    # Instead of: deno test
+deno task cache   # Instead of: deno cache --reload src/**/*.ts
+```
+
+**Manual Override**: For any other Deno command, prefix with the environment variable:
+```bash
+DENO_TLS_CA_STORE=system deno [your-command]
+```
+
+**Permanent Fix**: Add this to your shell profile (`~/.zshrc`):
+```bash
+export DENO_TLS_CA_STORE=system
+```
+
 # Usage
 
 If your project already relies on the `axios` library, `xmas` will just use it to send http requests
@@ -22,9 +59,19 @@ const xmas = new Xmas(config);
 
 // Create a new group in your xMatters instance:
 const group = { targetName: 'API developers' };
-xmas.groups.create(group)
-  .then(handleSuccess)
-  .catch(handleError);
+const response = await xmas.groups.save(group);
+
+// Access the HTTP response details:
+console.log('Status:', response.status);
+console.log('Headers:', response.headers);
+console.log('Created group:', response.body);
+
+// Get groups with pagination:
+const groupsResponse = await xmas.groups.getGroups({ limit: 10, offset: 0 });
+console.log('Total groups:', groupsResponse.body.total);
+groupsResponse.body.data.forEach(group => {
+  console.log('Group:', group.targetName);
+});
 ```
 
 Alternative `config` object for **OAuth** (say when you have already generated tokens and safely
